requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.18.0
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.9
jinja2>=3.1.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
mail-parser>=3.15.0
# 추가 의존성
python-pptx>=0.6.21
PyPDF2>=3.0.0
python-docx>=1.1.0
websockets>=12.0
